Refactor Vertica test file
